added tests, coverage configured and fixs
