--- .php_cs.dist
@@ -1,113 +1,0 @@
-<?php
-
-$header = <<<'EOF'
-PHP CS Fixer for Reyesoft projects
-EOF;
-
-return PhpCsFixer\Config::create()
-    ->setRiskyAllowed(true)
-    ->setRules([
-        '@Symfony' => true,
-        '@Symfony:risky' => true,
-        // '@PSR2' => true, // cool, but add break lines on every fucntion with 1+ params
-        '@PHPUnit60Migration:risky' => true,
-        'array_syntax' => ['syntax' => 'short'],
-        'combine_consecutive_unsets' => true,
-        'general_phpdoc_annotation_remove' => ['expectedException', 'expectedExceptionMessage', 'expectedExceptionMessageRegExp'],
-        'no_short_echo_tag' => true,
-        'no_unreachable_default_argument_value' => true,
-        'no_useless_return' => true,
-        'braces' => ['allow_single_line_closure' => true],
-        'ordered_imports' => true,
-        'no_unused_imports' => true,
-        'semicolon_after_instruction' => true,
-
-        /* added by pablorsk */
-        'ternary_to_null_coalescing' => true,
-        'binary_operator_spaces' => true,
-        'single_blank_line_before_namespace' => true,
-        'random_api_migration' => true,
-        'concat_space' => ['spacing' => 'one'],
-        'no_unreachable_default_argument_value' => true,
-        'class_definition' => ['singleLine' => true, 'singleItemSingleLine' => true],
-        'no_unreachable_default_argument_value' => true,
-
-        /* php-cs-fixer 2.7 */
-        'yoda_style' => false,
-        'combine_consecutive_issets' => true,
-        'no_homoglyph_names' => true,
-
-        /* php-cs-fixer 2.8/2.9 */
-        'method_separation' => true,
-        'explicit_indirect_variable' => true,
-        'phpdoc_align' => true,
-        'align_multiline_comment' => true,
-        'compact_nullable_typehint' => true,
-        'hash_to_slash_comment' => true,
-        'linebreak_after_opening_tag' => true,
-        'method_chaining_indentation' => true,
-
-        /* php-cs-fixer 2.10/2.11 */
-        'multiline_whitespace_before_semicolons' => true,
-        'no_alternative_syntax' => true,
-        'ordered_imports' => true,
-        'php_unit_ordered_covers' => true,
-        'php_unit_construct' => true,
-        // 'php_unit_strict' => true,   //  API-941
-        'string_line_ending' => true,
-        'php_unit_set_up_tear_down_visibility' => true,
-        'function_to_constant' => true,
-        'magic_constant_casing' => true,
-        'single_quote' => true,
-        'fully_qualified_strict_types' => true,
-        'date_time_immutable' => true,  // ver implicancias de este cambio
-        'array_indentation' => true,
-
-        /* PHP 7.0 */
-        '@PHP70Migration:risky' => true,
-
-        /* PHP 7.1 */
-        '@PHP71Migration' => true,
-        '@PHP71Migration:risky' => true,
-        'declare_strict_types' => true,
-
-        // on future?
-        'explicit_string_variable' => true, //check classes like $user->first_name
-        // 'strict_comparison' => true,
-        // 'no_useless_else' => true,
-        /*
-        'ordered_class_elements' => [
-            'use_trait',
-            'constant_public',
-            'constant_protected',
-            'constant_private',
-            'property_public',
-            'property_protected',
-            'property_private',
-            'construct',
-            'destruct',
-            'magic',
-            'phpunit',
-            'method_public',
-            'method_protected',
-            'method_private'
-        ],
-        */
-        'header_comment' => [
-            'header' =>
-                "This file is part of Multinexo PHP Afip WS package.\n".
-                "\n".
-                "Copyright (C) 1997-2018 Reyesoft <info@reyesoft.com>.\n".
-                "\n".
-                "For the full copyright and license information, please view the LICENSE",
-            'commentType' => 'PHPDoc',
-            'location' => 'after_open',
-            'separate' => 'bottom'
-        ],
-    ])
-    ->setFinder(
-        PhpCsFixer\Finder::create()
-            ->exclude('tests/Fixtures')
-            ->in(__DIR__)
-    )
-;
