+++ config/config.php
@@ -25,7 +25,7 @@ return [
     ],
 
     'archivos' => [
-        'wsaaWsdl' => __DIR__ . '/../Multinexo/Afip/WSAA/wsaa.wsdl',
+        'wsaaWsdl' => __DIR__ . '/../src/WSAA/wsaa.wsdl',
         'certificado' => null,
         'clavePrivada' => null,
     ],
