+++ phpunit.xml
@@ -14,6 +14,11 @@
          beStrictAboutTodoAnnotatedTests="true"
          processIsolation="false"
          stopOnFailure="true">
+    <logging>
+        <log type="coverage-html" target="resources/.tmp/reports/coverage" lowUpperBound="55" highLowerBound="85" />
+        <log type="coverage-clover" target="resources/.tmp/clover.xml"/>
+        <log type="testdox-text" target="resources/.tmp/reports/executed.txt"/>
+    </logging>
     <testsuites>
         <testsuite name="integration">
             <directory>./tests/</directory>
@@ -21,7 +26,7 @@
     </testsuites>
     <filter>
         <whitelist>
-            <directory suffix=".php">app/</directory>
+            <directory suffix=".php">src/</directory>
         </whitelist>
     </filter>
