+++ resources/ci/php-cpd.sh
@@ -1,4 +1,6 @@
 #!/bin/sh
 
-./vendor/bin/phpcpd --min-tokens=50 ./src/ ./routes/ \
+./vendor/bin/phpcpd --min-tokens=50 ./src/ \
+--regexps-exclude=ManejadorResultados.php \
+&&
 exit $?
