--- resources/ci/reset.sh
@@ -1,5 +1,0 @@
-#!/bin/sh
-
-sh resources/ci/pipeline/laravel-reset.sh &&
-
-exit $?
