--- tests/Afip/AfipTest.php
@@ -1,22 +1,0 @@
-<?php
-/**
- * Copyright (C) 1997-2018 Reyesoft <info@reyesoft.com>.
- *
- * This file is part of php-afip-ws. php-afip-ws can not be copied and/or
- * distributed without the express permission of Reyesoft
- */
-
-declare(strict_types=1);
-
-namespace Tests\Afip;
-
-/**
- * This class doesn't do much yet..
- */
-class AfipTest extends \PHPUnit\Framework\TestCase
-{
-    public function testisValid(): void
-    {
-        $this->assertTrue(true);
-    }
-}
